+++ HabboHotel/Items/Wired/Boxes/Conditions/ActorHasHandItemBox.cs
@@ -6,7 +6,7 @@ using Plus.HabboHotel.Users;
 
 namespace Plus.HabboHotel.Items.Wired.Boxes.Conditions;
 
-internal class ActorHasHandItemBox : IWiredItem, IWiredExecutable
+internal class ActorHasHandItemBox : IWiredItem, IWiredExecutable, IWiredActorExecutable
 {
     public ActorHasHandItemBox(Room instance, Item item)
     {
@@ -33,6 +33,11 @@ internal class ActorHasHandItemBox : IWiredItem, IWiredExecutable
     }
 
     bool IWiredExecutable.Execute(WiredExecutionContext context)
+    {
+        return ((IWiredActorExecutable)this).Execute((WiredActorExecutionContext)context);
+    }
+
+    bool IWiredActorExecutable.Execute(WiredActorExecutionContext context)
     {
         if (Instance == null || string.IsNullOrEmpty(StringData))
             return false;